+++ backend/courses/quizzes/api/admin/delete_quiz.py
@@ -1,13 +1,7 @@
 from rest_framework.decorators import api_view, permission_classes
 from rest_framework import permissions, status
-from courses.quizzes.api.admin.schema import CreateQuizRequestSerializer
 from rest_framework.response import Response
-from rest_framework import status
 import courses.models as db
-from django.db.models import Q
-from django.contrib.contenttypes.models import ContentType
-import requests
-from datetime import datetime
 
 
 @api_view(["DELETE"])
