+++ backend/courses/quizzes/api/admin/question/create/create_coding_question.py
@@ -2,23 +2,11 @@ from rest_framework.decorators import api_view, permission_classes
 from rest_framework import permissions, status
 from courses.quizzes.api.admin.schema import CreateCodingQuestionRequestSerializer
 from rest_framework.response import Response
-from django.db.models import Q
 import courses.models as db
-
-
-class CustomException(Exception):
-    pass
-
-
-def validate_user_is_ta_or_instructor_in_course(user_id: int, course_slug: str) -> None:
-    try:
-        db.Enrollment.objects.get(
-            (Q(role__kind=db.Role.Kind.TA) | Q(role__kind=db.Role.Kind.INSTRUCTOR)),
-            role__offering__course__slug=course_slug,
-            user_id=user_id,
-        )
-    except db.Enrollment.DoesNotExist:
-        raise CustomException("User is not a TA or Instructor in this course")
+from courses.quizzes.api.admin.utils import (
+    validate_user_is_ta_or_instructor_in_course,
+    CustomException,
+)
 
 
 @api_view(["POST"])
