+++ backend/courses/quizzes/api/admin/question/delete.py
@@ -1,8 +1,6 @@
 from rest_framework.decorators import api_view, permission_classes
 from rest_framework import permissions, status
-from courses.quizzes.api.admin.schema import CreateCheckboxQuestionRequestSerializer
 from rest_framework.response import Response
-from django.db.models import Q
 from uuid import UUID
 import courses.models as db
