+++ backend/courses/quizzes/api/admin/question/edit/edit_coding_question.py
@@ -2,7 +2,6 @@ from rest_framework.decorators import api_view, permission_classes
 from rest_framework import permissions, status
 from courses.quizzes.api.admin.schema import CreateCodingQuestionRequestSerializer
 from rest_framework.response import Response
-from django.db.models import Q
 import courses.models as db
