+++ backend/courses/quizzes/api/admin/question/edit/edit_written_response_question.py
@@ -4,7 +4,6 @@ from courses.quizzes.api.admin.schema import (
     CreateWrittenResponseQuestionRequestSerializer,
 )
 from rest_framework.response import Response
-from django.db.models import Q
 import courses.models as db
