+++ backend/courses/quizzes/api/complete.py
@@ -1,5 +1,3 @@
-import courses.models as db
-from uuid import UUID
 from rest_framework.response import Response
 from rest_framework import status
 from rest_framework import permissions
