+++ backend/courses/quizzes/api/utils.py
@@ -4,7 +4,6 @@ from rest_framework.response import Response
 from rest_framework import status
 from datetime import datetime
 from typing import Optional, Union, Type
-from rest_framework.exceptions import ValidationError
 from django.utils import timezone
